pyproject.toml
[tool.poetry]
name = "innoter"
version = "0.1.0"
description = ""
authors = ["KateM_zz <zupperu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
isort = "^5.13.2"
black = "^24.1.0"
pytest = "^7.4.4"
ruff = "^0.1.14"
mysqlclient = "^2.2.1"
django-environ = "^0.11.2"


[tool.black]
target-version = ['py311']
include = '\.pyi?$'
line-length = 100

[tool.isort]
profile = "black"
filter_files = true
line_length = 100

[tool.ruff]
line-length = 100
exclude = [
    "**/migrations/*.py"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"